Fix the generate_coverage.js script
